python/payment-orchestrator/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
confluent-kafka
